refactor: extract functions into separate files
